% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_summarize.R
\name{col_summarize}
\alias{col_summarize}
\title{The col_summarize function}
\usage{
col_summarize(MAT, dig = 2, level = 0.95)
}
\arguments{
\item{MAT}{A matrix}

\item{dig}{Number of digits to round estimates and CrIs to}

\item{level}{Confidence level}
}
\description{
Function to quickly return credible intervals
}
\examples{
M <- matrix(rnorm(1000),ncol = 4)
col_summarize(M)
}
\keyword{Credible}
\keyword{intervals}
