% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1_data.R
\docType{data}
\name{example1_data}
\alias{example1_data}
\title{Simulated MSN example data.}
\format{
A list of 3:
\describe{
  \item{Y}{a 1000x4 outcome matrix}
  \item{X}{A 1000x2 predictor matrix}
  \item{W}{A 1000x2 predictor matrix}
}
}
\usage{
data(example1_data)
}
\description{
A dataset containing Y, X, and W for n = 1000, p = 2, v = 2, and K = 3.
}
\keyword{datasets}
