% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_msn.R
\name{fit_msn}
\alias{fit_msn}
\title{Fit MSN mixture model}
\usage{
fit_msn(Y, X, W, K, nsim, burn)
}
\arguments{
\item{Y}{matrix of non-missing outcome data}

\item{X}{matrix of predictors used in the MSN model component}

\item{W}{matrix of predictors used in the multinomial logit clustering component}

\item{K}{number of clusters to fit. Use WAIC to choose best K.}

\item{nsim}{number of MCMC iterations to perform}

\item{burn}{number of initial MCMC iterations to discard. The function will return nsim - burn total posterior samples.}
}
\description{
This fits the MSN mixture model without imputation
}
\examples{
data(example1_data)
fit1 <- fit_msn(Y = example1_data$Y,
                X = example1_data$X,
                W = example1_data$W,
                K = 3,
                nsim = 10,
                burn = 0)
}
\keyword{Bayesian,}
\keyword{MSN,}
\keyword{clustering}
\keyword{mixture}
\keyword{model,}
