% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_msn_impute.R
\name{fit_msn_impute}
\alias{fit_msn_impute}
\title{Fit MSN mixture model}
\usage{
fit_msn_impute(Y, X, W, K, Q, nsim = 1000, burn = 0, kmeans_init = TRUE)
}
\arguments{
\item{Y}{matrix of non-missing outcome data}

\item{X}{matrix of predictors used in the MSN model component}

\item{W}{matrix of predictors used in the multinomial logit clustering component}

\item{K}{number of clusters to fit. Use WAIC to choose best K.}

\item{Q}{matrix of predictors used in the missing data model.}

\item{nsim}{number of MCMC iterations to perform}

\item{burn}{number of initial MCMC iterations to discard. The function will return nsim - burn total posterior samples.}

\item{kmeans_init}{logical indicating if multivariate k-means should be used to initialize clusters.}
}
\description{
This fits the MSN mixture model with imputation
}
\examples{
data(example2_data)
fit2 <- fit_msn_impute(Y = example2_data$Y,
                X = example2_data$X,
                W = example2_data$W,
                Q = example2_data$Q,
                K = 3,
                nsim = 10,
                burn = 0)
}
\keyword{Bayesian,}
\keyword{MSN,}
\keyword{clustering}
\keyword{mixture}
\keyword{model,}
