% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mst.R
\name{fit_mst}
\alias{fit_mst}
\title{Fit MST mixture model}
\usage{
fit_mst(Y, X, W, K, nu = 3, nsim = 1000, burn = 0)
}
\arguments{
\item{Y}{matrix of non-missing outcome data}

\item{X}{matrix of predictors used in the MSN model component}

\item{W}{matrix of predictors used in the multinomial logit clustering component}

\item{K}{number of clusters to fit. Use WAIC to choose best K.}

\item{nu}{degrees of freedom parameter for skew-t model.}

\item{nsim}{number of MCMC iterations to perform}

\item{burn}{number of initial MCMC iterations to discard. The function will return nsim - burn total posterior samples.}
}
\description{
This function fits the MST mixture model without imputation
}
\examples{
data(example1_data)
fit1 <- fit_mst(Y = example1_data$Y,
                X = example1_data$X,
                W = example1_data$W,
                K = 3,
                nsim = 10,
                burn = 0)
}
\keyword{Bayesian,}
\keyword{MSN,}
\keyword{clustering}
\keyword{mixture}
\keyword{model,}
