% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_CRI.R
\name{mean_CRI}
\alias{mean_CRI}
\title{The mean_CRI function}
\usage{
mean_CRI(y, dig = 2)
}
\arguments{
\item{y}{A numeric vector}

\item{dig}{The number of digits to round to}
}
\description{
Simple function to return the mean (95% CrI) for a vector
}
\examples{
mean_CRI(rnorm(1000))
}
\keyword{MCMC}
