% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{Compute WAIC for fitted MSN model}
\usage{
waic(fit)
}
\arguments{
\item{fit}{object returned by fit_msn()}
}
\description{
This computes WAIC for fitted MSN models to be used in model selection.
}
\examples{
data(example1_data)
fit1 <- fit_msn(Y = example1_data$Y,
                X = example1_data$X,
                W = example1_data$W,
                K = 3,
                nsim = 100,
                burn = 0)
waic(fit1)
}
\keyword{Bayesian,}
\keyword{MSN,}
\keyword{clustering}
\keyword{mixture}
\keyword{model,}
